Reorganize sample Storybook components and pages
Turns out I have an opinion about where components and pages should live
after all! This commit organizes the Storybook ones, and a future commit
will organize our own application.

- Move non-page components to `/src/components`
- Move pages (things that will be routed to) to `/src/pages`
- In both cases, further nest these Storybook examples inside `example/`
  directories
- Keep styles alongside their components

Issue #10 Scaffold routing and page construction
